+++ string/sha1sum.hpp
@@ -32,7 +32,7 @@ class SHA1 {
     void transform(uint32 block[BLOCK_BYTES]);
 
     static void buffer_to_block(const std::string& buffer, uint32 block[BLOCK_BYTES]);
-    static void read(std::istream& is, std::string& s, int max);
+    static void read(std::istream& is, std::string& s, size_t max);
 };
 
 std::string sha1(const std::string& string);
